FIX: updated sorting algorithm, Points = (clicks + 1) * exp(c1 * distance) * exp(c2 * time)
c1 and c2 are negative constants.
Distance is the distance between the current location and the picture
time the time between the current time and the time the picture was taken.
Clicks is the amount of "so berlin" votes the photo has received

The constants are:
c1 = -7e-4
c2 = -1.15e-09
Assuming distance in km for c1 and milliseconds for c2. As discussed, we might have to readjust it a bit in the beta phase, but i am sure that after that we don't have to change it again.